api/requirements.txt
# Flask
Flask==3.0.3

# Kafka core
kafka-python==2.2.15

# MongoDB driver
pymongo==4.7.2

# Dotenv para carregar variáveis de ambiente
python-dotenv==1.0.1

# Docker para integração com o Docker
docker

# Requests para fazer requisições HTTP
requests

# Google Generative AI para integração com a API
google-generativeai